Add a config setting to perform strict checks of return statements

+++ src/Phan/LanguageServer/Protocol/Position.php
@@ -67,6 +67,7 @@ class Position
     {
         $lines = explode("\n", $content);
         $slice = array_slice($lines, 0, $this->line);
-        return array_sum(array_map('strlen', $slice)) + count($slice) + $this->character;
+        // TODO: array_sum should infer sum of ints is typically an int
+        return ((int)array_sum(array_map('strlen', $slice))) + count($slice) + $this->character;
     }
 }
